refactor(krt): move status_line to the live module
A move of the function and of the two tests that read its line. The line stays the same, and the run loop calls the module that owns the display.
